+++ src/main/java/subway/ui/ExceptionController.java
@@ -4,7 +4,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-import subway.dto.ExceptionResponse;
+import subway.ui.dto.response.ExceptionResponse;
 
 import java.sql.SQLException;
