+++ src/main/java/subway/ui/dto/request/CreationStationRequest.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.ui.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
