+++ src/main/java/subway/ui/dto/response/ExceptionResponse.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.ui.dto.response;
 
 public class ExceptionResponse {
