+++ src/main/java/subway/ui/dto/response/ReadLineResponse.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.ui.dto.response;
 
 import subway.domain.Line;
