+++ src/main/java/subway/ui/dto/response/ReadStationResponse.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.ui.dto.response;
 
 import subway.domain.Station;
