+++ src/test/java/subway/integration/StationIntegrationTest.java
@@ -10,8 +10,8 @@ import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import subway.dto.CreationStationRequest;
-import subway.dto.ReadStationResponse;
+import subway.ui.dto.request.CreationStationRequest;
+import subway.ui.dto.response.ReadStationResponse;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -23,7 +23,7 @@ public class StationIntegrationTest extends IntegrationTest {
     @Test
     void 지하철역을_생성한다() {
         // given
-        final CreationStationRequest requestDto =  CreationStationRequest.from("강남역");
+        final CreationStationRequest requestDto = CreationStationRequest.from("강남역");
 
         // when
         final ExtractableResponse<Response> response = RestAssured.given().log().all()
@@ -44,7 +44,7 @@ public class StationIntegrationTest extends IntegrationTest {
     @Test
     void 기존에_존재하는_지하철역_이름으로_지하철역을_생성하면_상태코드_400을_반환한다() {
         // given
-        final CreationStationRequest requestDto =  CreationStationRequest.from("양재역");
+        final CreationStationRequest requestDto = CreationStationRequest.from("양재역");
 
         RestAssured.given().log().all()
                 .body(requestDto)
@@ -71,7 +71,7 @@ public class StationIntegrationTest extends IntegrationTest {
     @Test
     void 지하철역을_조회한다() {
         /// given
-        final CreationStationRequest requestDto =  CreationStationRequest.from("선릉역");
+        final CreationStationRequest requestDto = CreationStationRequest.from("선릉역");
         final ExtractableResponse<Response> createResponse = RestAssured.given().log().all()
                 .body(requestDto)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
@@ -99,7 +99,7 @@ public class StationIntegrationTest extends IntegrationTest {
     @Test
     void 지하철역을_제거한다() {
         // given
-        final CreationStationRequest requestDto =  CreationStationRequest.from("잠실역");
+        final CreationStationRequest requestDto = CreationStationRequest.from("잠실역");
         final ExtractableResponse<Response> createResponse = RestAssured.given().log().all()
                 .body(requestDto)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
